refactor(manager): apply review
