+++ packages/manager/src/components/ui/PathCard.tsx
@@ -1,23 +1,20 @@
-import { AssetStat } from "@assetbox/tools";
-import React from "react";
-
 import { cn } from "../../utils";
 
-interface PathCardProps {
-  type: AssetStat["type"];
+interface PathCardProps extends React.HTMLAttributes<HTMLDivElement> {
   paths: string[];
 }
 
-export const PathCard = ({ paths, type }: PathCardProps) => {
+export const PathCard = ({ paths, className, ...props }: PathCardProps) => {
   return (
     <div
       className={cn(
         "grow bg-[#F7F9Fb] rounded-lg p-6 overflow-y-scroll",
-        type === "image" && "flex-1"
+        className
       )}
+      {...props}
     >
       {paths.map((path, index) => (
-        <div className="text-gray-dark ">
+        <div className="text-gray-dark" key={`path-${path}`}>
           <span className="text-black text-opacity-20 mr-6">{index + 1}</span>
           {path}
         </div>
